Parquet resolver to speed up resume or extending the dataset
