+++ datasets_preparation/prepare_recipe_data.py
@@ -1,6 +1,5 @@
 from logger import logger
 from config import TrainingStage
-from datasets_preparation.data_preparation_utils import get_max_number_of_cpu_processes
 from datasets_preparation.evals.prepare_hellaswag_dataset import prepare_hellaswag_dataset
 from datasets_preparation.evals.prepare_winogrande_dataset import prepare_winogrande_dataset
 from datasets_preparation.evals.prepare_arc_challenge_dataset import prepare_arc_challenge_dataset
